Data Structures and Algorithms 2.2
Create a data file consisting of at least 100 string values. You can
create the list yourself, or perhaps copy the values from a text file of
some type, or you can even create the file by generating random strings.
Sort the file using each of the sorting algorithms mentioned in 2.1
above. There are a number of things that can go wrong when a program
accesses an external resource such as a file. E.g. file may not exit in
the folder the program expects it to be, file may be in a different
format etc. Your program should handle these and other exceptions to
demonstrate the implementation of error handling and reporting.